refactor: move stuff out of __main__
